refactored tileTypes
