+++ src/app/features/game/models/tile.ts
@@ -1,9 +1,9 @@
-import { TileType } from "./types";
+import { TileType, TileTypeName } from "./types";
 
 export class Tile {
-    public type: TileType;
+    public type;
     public value: string = " ";
-    
+
     public constructor(typeAndVal: string = "$e") {
         if (typeAndVal.length < 2)
             typeAndVal = "$e";
@@ -11,12 +11,12 @@ export class Tile {
         this.type = this.typeFromCode(typeAndVal.at(0)!);
         this.value = typeAndVal.at(1)!;
     }
-    
-    public typeFromCode(code: string): TileType {
+
+    public typeFromCode(code: string) {
         switch (code) {
-            case "'": return TileType.GROUND;
-            case "|": return TileType.WALL;
-            case "-": return TileType.DANGER;
+            case TileType.GROUND.charCode: return TileType.GROUND;
+            case TileType.WALL.charCode: return TileType.WALL;
+            case TileType.DANGER.charCode: return TileType.DANGER;
             default: return TileType.EMPTY;
         }
     }
