+++ src/app/features/game/services/input/input-router.service.ts
@@ -4,7 +4,7 @@ import { CommandParser } from "./command-parser.service";
 import { GameState } from "../game-state.service";
 import { Injectable } from "@angular/core";
 import { fromEvent } from "rxjs";
-import { WriterService } from "./writer.service";
+import { WriterService } from "../writer.service";
 
 @Injectable()
 export class InputRouter {
